Moved source code to software/app
And renamed to my_app.*
