--- CatCutifier/CMakeLists.txt
@@ -1,4 +1,0 @@
-﻿add_library (CatCutifier "CatCutifier.cpp" "CatCutifier.h")
-target_include_directories(CatCutifier PUBLIC .)
-
-# Should install library properly here, but I'm focusing on just the docs
