+++ software/CMakeLists.txt
@@ -1,0 +1,4 @@
+﻿add_library (CatCutifier "app/my_app.cpp" "app/my_app.h")
+target_include_directories(CatCutifier PUBLIC app)
+
+# Should install library properly here, but I'm focusing on just the docs
