+++ software/app/my_app.cpp
@@ -1,4 +1,4 @@
-#include "CatCutifier.h"
+#include "my_app.h"
 
 void cat::make_cute() {
 	// Magic happens
